// JcCar-client/src/app/shared/services/model/chatEvaluationComprehension.ts
/**
 * ALERTA API
 * SMS - Módulo Alertas de SMS que tem como objetivo comunicar aos interessados deliberações das
 * Agências Reguladoras e eventos de segurança relevantes para que as medidas necessárias sejam tomadas
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Alert } from './alert';
import { AlertInvolved } from './alertInvolved';
import { ComprehensiveEvaluation } from './comprehensiveEvaluation';


export interface ChatEvaluationComprehension {
    alert?: Alert;
    alertInvolved?: AlertInvolved;
    evaluationComprehension?: ComprehensiveEvaluation;
    id?: number;
    message?: string;
    receiver?: string;
    senDateMassage?: Date;
    sender?: string;
}